requirements.txt
numpy==1.24.4
mysql-connector-python
matplotlib

